+++ frontend/app/app.vue
@@ -6,7 +6,6 @@ import { fr } from "@nuxt/ui/locale";
 const colorMode = useColorMode();
 colorMode.value = "dark";
 colorMode.preference = "dark";
-
 </script>
 
 <template>
