+++ frontend/app/assets/css/main.css
@@ -54,10 +54,9 @@ p {
     --color-slate-300: #90a1b9;
     --color-slate-200: #c8dbe8;
 
-
     --color-dark-950: #1a1f2e;
     --color-dark-900: #232a3b;
-    --color-dark-850: #202D43;
+    --color-dark-850: #202d43;
     --color-dark-800: #2e3750;
     --color-dark-700: #3d4a65;
     --color-dark-600: #5a6a8a;
