+++ frontend/app/components/home/ImportantInformationSection/ImportantInformationSection.vue
