+++ frontend/app/components/home/Last30DaysSection/LastMonthSection.vue
