+++ frontend/app/components/home/Last365DaysSection/LastYearSection.vue
