+++ frontend/app/components/layout/HomeLayout.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import ImportantInformationSection from "../home/ImportantInformationSection/index.vue";
-import TodaySection from "../home/TodaySection/index.vue";
-import LastMonthSection from "../home/Last30DaysSection/index.vue";
-import LastYearSection from "../home/Last365DaysSection/index.vue";
+import ImportantInformationSection from "../home/ImportantInformationSection/ImportantInformationSection.vue";
+import TodaySection from "../home/TodaySection/TodaySection.vue";
+import LastMonthSection from "../home/Last30DaysSection/LastMonthSection.vue";
+import LastYearSection from "../home/Last365DaysSection/LastYearSection.vue";
 </script>
 <template>
     <div class="py-8">
